Module sampitor::view::filter

Components for fitlering signals.

Re-exports

pub use base::Filter;

pub use base::Knob;

pub use normalize::Normalize;

Modules

base
Fundemental traits for digital signal processing user interface components.

normalize
User interface for changing amplitude levels of a signal.

Structs

Filters
A UI view for navigating the file system, reading audio files, and writing audio files.